Create README.md
Enthusiastic, results-driven, and accomplished geophysicist with nearly fifteen years of experience in qualitative and quantitative interpretation of geoscientific data with several onshore and offshore exploration, appraisal and development projects. Competent in plan and implement most appropriate workflows to address geophysical challenges in seismic data interpretation, seismic inversion, rock physics analysis and reservoir modeling. Proven capability to manage multiple projects simultaneously and meet crucial deadlines as per the compliance with regulatory requirements and business strategy. Adaptable and transformational professional with an ability to work independently or as part of a multidisciplinary team of geologists, petrophysicists and reservoir engineers.